Moving the example plugin.
