+++ example/plugins/example.js
